Address PR feedback for #4201: migration head, unused imports, update ordering, aiosqlite, relocated test imports
