+++ pyproject.toml
@@ -48,6 +48,7 @@ pytest-mock = { version = "~=3.15.1", optional = true }
 pytest-asyncio = { version = "^1.3.0", optional = true }
 pytest-xdist = { version = "~=3.8.0", optional = true }
 pytest-timeout = { version = "~=2.4.0", optional = true }
+aiosqlite = { version = ">=0.19.0,<0.22.0", optional = true }
 typer = ">=0.23.1,<0.28.0"
 click = "8.5.0"
 gitpython = "^3.1.46"
@@ -58,7 +59,7 @@ opentelemetry-exporter-otlp = "^1.25.0"
 opentelemetry-instrumentation-fastapi = ">=0.46b0,<1.0"
 
 [tool.poetry.extras]
-tests = ["pytest", "pytest-mock", "pytest-asyncio", "pytest-xdist", "pytest-timeout"]
+tests = ["pytest", "pytest-mock", "pytest-asyncio", "pytest-xdist", "pytest-timeout", "aiosqlite"]
 
 [tool.poetry.scripts]
 moonmind = "moonmind.cli:main"
