+++ tests/unit/api/routers/test_omnigent_bridge.py
@@ -29,7 +29,7 @@ from api_service.api.routers.omnigent_bridge import (
     embedded_host_auth_preflight,
     router,
 )
-from api_service.api.routers.retrieval_gateway import get_capability_registry
+from api_service.api.routers.omnigent_bridge import get_capability_registry
 from api_service.auth_providers import get_current_user
 from moonmind.omnigent.bridge_config import (
     HOST_PROTOCOL_MODE_EMBEDDED,
