+++ tests/unit/api/routers/test_omnigent_workflow_chat.py
@@ -34,7 +34,7 @@ from api_service.api.routers.omnigent_bridge import (
     _validate_native_resource_path,
     workflow_chat_router,
 )
-from api_service.api.routers.retrieval_gateway import get_capability_registry
+from api_service.api.routers.omnigent_bridge import get_capability_registry
 from api_service.auth_providers import get_current_user
 from moonmind.omnigent import native_ui_compat
 from moonmind.omnigent.bridge_config import (
